server/pyproject.toml
[tool.poetry]
name = "under-stress"
version = "0.1.0"
description = ""
authors = ["Ihor <cguilliman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
Django = "^4.2.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
